Refactor node logic into module
